Move db connection to a connections folder.
